refactor: tighten cloud webdriver provider internals
