+++ src/cloud-webdriver/webdriver-gestures.ts
@@ -1,4 +1,4 @@
-import { centerOfRect } from '../kernel/snapshot.ts';
+import { centerOfRect, type Rect } from '../kernel/snapshot.ts';
 import type { ScrollDirection } from '../core/scroll-gesture.ts';
 import type { W3CActionSequence } from './webdriver-client.ts';
 
@@ -14,8 +14,8 @@ export function touchPointer(name: string, actions: Record<string, unknown>[]): 
 export function scrollStart(
   direction: ScrollDirection,
   distance: number,
+  rect: Rect,
 ): { x: number; y: number } {
-  const rect = { x: 0, y: 0, width: 400, height: 800 };
   const center = centerOfRect(rect);
   switch (direction) {
     case 'up':
@@ -29,8 +29,12 @@ export function scrollStart(
   }
 }
 
-export function scrollEnd(direction: ScrollDirection, distance: number): { x: number; y: number } {
-  const start = scrollStart(direction, distance);
+export function scrollEnd(
+  direction: ScrollDirection,
+  distance: number,
+  rect: Rect,
+): { x: number; y: number } {
+  const start = scrollStart(direction, distance, rect);
   switch (direction) {
     case 'up':
       return { ...start, y: start.y - distance };
