+++ src/cloud-webdriver/webdriver-utils.ts
@@ -1,0 +1,29 @@
+import type { DeviceLease } from '../daemon/lease-registry.ts';
+
+export type LeaseValue<T> = T | ((lease: DeviceLease) => T);
+
+export function resolveLeaseValue<T>(
+  value: LeaseValue<T> | undefined,
+  lease: DeviceLease,
+): T | undefined {
+  return typeof value === 'function' ? (value as (lease: DeviceLease) => T)(lease) : value;
+}
+
+export function basicAuthHeader(credentials: { username: string; accessKey: string }): string {
+  return `Basic ${Buffer.from(`${credentials.username}:${credentials.accessKey}`).toString('base64')}`;
+}
+
+export function trimLeadingSlash(value: string): string {
+  return value.replace(/^\/+/, '');
+}
+
+export function trimTrailingSlash(value: string): string {
+  return value.replace(/\/+$/, '');
+}
+
+export function withTrailingSlash(url: URL): URL {
+  if (url.pathname.endsWith('/')) return url;
+  const copy = new URL(url);
+  copy.pathname = `${copy.pathname}/`;
+  return copy;
+}
